requirements.txt
google-genai>=1.0.0
openai>=1.0.0
pyyaml>=6.0
rich>=13.0.0
python-dotenv>=1.0.0
flask>=3.0.0
weasyprint>=62.0
